Group each composable by sections into its own Kotlin file  * The sections currently are SearchBar, NewsFeed, and BottomBar
